+++ test-app/app/serializers/AlternateDbConnectionPostSerializer.ts
@@ -1,0 +1,9 @@
+import { DreamSerializer } from '../../../src/index.js'
+import AlternateDbConnectionPost from '../models/AlternateDbConnectionPost.js'
+
+export const AlternateDbConnectionPostSummarySerializer = (
+  alternateDbConnectionPost: AlternateDbConnectionPost
+) => DreamSerializer(AlternateDbConnectionPost, alternateDbConnectionPost).attribute('id')
+
+export const AlternateDbConnectionPostSerializer = (alternateDbConnectionPost: AlternateDbConnectionPost) =>
+  AlternateDbConnectionPostSummarySerializer(alternateDbConnectionPost).attribute('body')