+++ test-app/db/migrations/1682458327691-test-manual-new-transaction-2.ts
@@ -1,0 +1,16 @@
+import { Kysely, sql } from 'kysely'
+import { DreamMigrationHelpers } from '../../../src/index.js'
+
+export async function up(db: Kysely<any>): Promise<void> {
+  DreamMigrationHelpers.newTransaction()
+
+  await db.schema
+    .alterTable('pets')
+    .addCheckConstraint(
+      'species_migration_transaction_test_constraint',
+      sql`species != 'migration_transaction_test'`
+    )
+    .execute()
+}
+
+export async function down(): Promise<void> {}