+++ CHANGELOG.md
@@ -32,7 +32,8 @@ yarn psy sync
 
 import Dream from '../../../src/Dream.js'
 import { DBClass } from '../../types/db.alternateConnection.js'
-import { dreamTypeConfig, schema } from '../../types/dream.alternateConnection.js'
+import { connectionTypeConfig, schema } from '../../types/dream.alternateConnection.js'
+import { globalTypeConfig } from '../../types/dream.globals.js'
 
 export default class MyAlternateConnectionApplicationModel extends Dream {
   declare public DB: DBClass
@@ -45,8 +46,12 @@ export default class MyAlternateConnectionApplicationModel extends Dream {
     return schema
   }
 
-  public override get dreamTypeConfig() {
-    return dreamTypeConfig
+  public override get connectionTypeConfig() {
+    return connectionTypeConfig
+  }
+
+  public override get globalTypeConfig() {
+    return globalTypeConfig
   }
 }
 ```