+++ CHANGELOG.md
@@ -32,7 +32,7 @@ yarn psy sync
 
 import Dream from '../../../src/Dream.js'
 import { DBClass } from '../../types/db.alternateConnection.js'
-import { globalSchema, schema } from '../../types/dream.alternateConnection.js'
+import { dreamTypeConfig, schema } from '../../types/dream.alternateConnection.js'
 
 export default class MyAlternateConnectionApplicationModel extends Dream {
   declare public DB: DBClass
@@ -45,8 +45,8 @@ export default class MyAlternateConnectionApplicationModel extends Dream {
     return schema
   }
 
-  public override get globalSchema() {
-    return globalSchema
+  public override get dreamTypeConfig() {
+    return dreamTypeConfig
   }
 }
 ```